test(checker): split cross file direct alias tests
